+++ sensu_plugin/__init__.py
@@ -2,6 +2,7 @@
 from sensu_plugin.plugin import SensuPlugin
 from sensu_plugin.check import SensuPluginCheck
 from sensu_plugin.metric import SensuPluginMetricJSON
+from sensu_plugin.metric import SensuPluginMetricInfluxdb
 from sensu_plugin.metric import SensuPluginMetricGraphite
 from sensu_plugin.metric import SensuPluginMetricStatsd
 from sensu_plugin.handler import SensuHandler